modification of ids (int to long)

+++ ProlineStudio-ResultExplorer/src/main/java/fr/proline/studio/rsmexplorer/gui/dialog/ImportIdentificationDialog.java
@@ -737,13 +737,13 @@ public class ImportIdentificationDialog extends DefaultDialog {
         return parameterList.getValues();
     }
 
-    public Integer getInstrumentId() {
+    public long getInstrumentId() {
         
         InstrumentConfiguration instrument = (InstrumentConfiguration) sourceParameterList.getParameter("instrument").getObjectValue(); 
         return instrument.getId();
     }
     
-    public Integer getPeaklistSoftwareId() {
+    public long getPeaklistSoftwareId() {
         PeaklistSoftware peaklistSoftware = (PeaklistSoftware) sourceParameterList.getParameter("peaklist_software").getObjectValue();  
         return peaklistSoftware.getId();
     }
